requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# Machine learning and AI
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0

# Image processing and computer vision
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.9.0

# Statistical analysis
statsmodels>=0.13.0

# Jupyter notebook support (for analysis)
jupyter>=1.0.0
ipykernel>=6.0.0

# Data visualization
plotly>=5.0.0

# Medical image analysis
nibabel>=3.2.0

# Utilities
tqdm>=4.60.0
argparse

# For reproducibility
setuptools>=50.0.0 